requirements.txt
--index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
torchvision==0.21.0+cu124
torchaudio==2.6.0+cu124
transformers==5.1.0
accelerate==1.12.0
bitsandbytes==0.49.1
qwen-vl-utils==0.0.14
pillow==12.0.0
pandas
openpyxl
huggingface_hub
